Небольшие изменения до начала второй лабы
