+++ AviaCompany/AviaCompany.Domain/Models/AircraftFamily.cs
@@ -10,7 +10,7 @@ public class AircraftFamily
     /// <summary>
     /// Уникальный идентификатор
     /// </summary>
-    public  int Id { get; set; }
+    public required int Id { get; set; }
 
     /// <summary>
     /// Название семейства
